J-114: * changed on useController
