+++ frontend/src/components/common/common.ts
@@ -14,7 +14,7 @@ export { default as Loader } from './loader/loader';
 export { default as Select } from './select/select';
 export { default as Datepiker } from './datepicker/datepicker';
 export { default as Table } from './table/table';
-export { default as InputImage } from './input-image/input-image';
+export { default as ImagePreviewControl } from './image-preview-control/image-preview-control';
 export { default as PodcastList } from './podcasts-list/podcasts-list';
 export { default as PodcastItem } from './podcast-item/podcast-item';
 export { default as Player } from './player/player';
