+++ frontend/src/components/common/image-preview-control/styles.module.scss
@@ -1,0 +1,34 @@
+.inputWrapper{
+  position: relative;
+  max-width: 716px;
+  width: 100%;
+  cursor: pointer;
+
+  &::before{
+    content: '';
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    height: 54px;
+    background: rgba(25, 34, 47, 0.7);
+    border-radius: 0 0 5px 5px;
+  }
+
+  &::after{
+    content: '';
+    position: absolute;
+    left: 50%;
+    bottom: 16px;
+    transform: translateX(-50%);
+    width: 24px;
+    height: 20px;
+    background: url('../../../assets/img/icon-camera.svg');
+    margin: auto;
+  }
+  
+  img{
+    border-radius: 5px;
+    height: 281px;
+    object-fit: cover;
+  }
+}
