+++ frontend/src/components/configurate-podcast/styles.module.scss
@@ -4,9 +4,9 @@
   max-width: 716px;
   width: 100%;
   margin: 0 auto;
-  padding: 0;
+  padding: 0 30px;
 
-  h2{
+  .podcastTitle{
     margin-top: 37px;
   }
 }
